Added edit function
Added the edit function. Also created spans to store the ID of a note being edited and to store the mode that the form is in (currently "create" and "edit" are used.

Changed the form type to post instead of submit.

Got "preventDefault" to work on the form, by putting it up at the top of the function and passing the event through the function. (see "e" in the saveNote function).

Added some commenting, need to add more. Cleaned up some old console logging and comments.